requirements.txt
fastapi
uvicorn
openai
pandas
scikit-learn
python-dotenv
streamlit
requests
deep-translator
sentence-transformers
faiss-cpu
python-multipart
